requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-groq
langchain-community
tavily-python
python-dotenv
pydantic>=2.0
rich
